Moved Limiter class to Client
